perverted 2-panels
Added index_3panels.html to retain integrity of original index.html
file. But still shares the same .scss file. So I guess not much effect
anyways.

Anyway, in this committ, I'm trying to expand to the 3rd panel on top
of the 2 exisiting ones (without modifying them).

The problem seems to be with the javascript to switch between the
id/classes. Content of 2 ids will apear on top of another, instead of
switching.

Possible solution:
- look into javascript
- look into the naming of IDs and Classes. Make sure they are properly
applied in the javascript.
- reorganize structure of country listings (wrapper), and content
showcasae (panel)