horizontal panel v0.1
working on the horizontal panels. now reduced to 2. trying to make
sense of the sizes of the two panels, as well as transitions between
them. the current easing works, but still seems a little abrupt.

also, the sizes of the list/content panels varies -- it should stay
fixed until break-points are met. otherwise the content becames really
hard to predict and manage.